requirements.txt
django 
djangorestframework
psycopg2-binary
python-decouple
djangorestframework-simplejwt
celery
celery[redis]